aria-compliance/requirements.txt
# Runtime dependencies — installed by Railway
anthropic>=0.40.0
voyageai>=0.3.0
neo4j>=5.0.0
supabase>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.27.0
